reimplement nft functionality
