+++ applications/tari_indexer/src/event_manager.rs
@@ -21,146 +21,191 @@
 //  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 use std::{
-    collections::{BTreeMap, HashMap, HashSet},
-    str::FromStr,
+    collections::BTreeMap,
+    str::FromStr, sync::Arc,
 };
 
-use futures::StreamExt;
 use log::*;
-use rand::{prelude::SliceRandom, rngs::OsRng};
-use tari_bor::decode;
-use tari_common::configuration::Network;
 use tari_crypto::tari_utilities::message_format::MessageFormat;
-use tari_dan_common_types::{committee::Committee, shard::Shard, Epoch, PeerAddress};
-use tari_dan_p2p::proto::rpc::{GetTransactionResultRequest, PayloadResultStatus, SyncBlocksRequest};
-use tari_dan_storage::consensus_models::{Block, BlockId, Command, Decision, TransactionRecord};
-use tari_engine_types::{commit_result::{ExecuteResult, TransactionResult}, events::Event, substate::{Substate, SubstateId}};
-use tari_epoch_manager::EpochManagerReader;
-use tari_template_lib::{models::Metadata, Hash};
-use tari_transaction::{Transaction, TransactionId};
-use tari_validator_node_rpc::client::{TariValidatorNodeRpcClientFactory, ValidatorNodeClientFactory};
+use tari_dan_app_utilities::substate_file_cache::SubstateFileCache;
+use tari_dan_common_types::PeerAddress;
+use tari_engine_types::{events::Event, substate::SubstateId};
+use tari_indexer_lib::substate_scanner::SubstateScanner;
+use tari_template_lib::{models::{Metadata, TemplateAddress}, Hash};
+use tari_transaction::TransactionId;
+use tari_validator_node_rpc::client::TariValidatorNodeRpcClientFactory;
+use tari_epoch_manager::base_layer::EpochManagerHandle;
 
-use crate::{config::EventFilterConfig, event_data::EventData, substate_storage_sqlite::{
-    models::{events::{NewEvent, NewScannedBlockId}, substate::NewSubstate},
+use crate::substate_storage_sqlite::{
+    models::events::NewEvent,
     sqlite_substate_store_factory::{
         SqliteSubstateStore,
         SubstateStore,
         SubstateStoreReadTransaction,
         SubstateStoreWriteTransaction,
     },
-}};
+};
 
 const LOG_TARGET: &str = "tari::indexer::event_manager";
 
 pub struct EventManager {
-    network: Network,
-    epoch_manager: Box<dyn EpochManagerReader<Addr = PeerAddress>>,
-    client_factory: TariValidatorNodeRpcClientFactory,
     substate_store: SqliteSubstateStore,
-    event_filters: Vec<EventFilterConfig>,
+    substate_scanner:
+        Arc<SubstateScanner<EpochManagerHandle<PeerAddress>, TariValidatorNodeRpcClientFactory, SubstateFileCache>>,
 }
 
 impl EventManager {
     pub fn new(
-        network: Network,
-        epoch_manager: Box<dyn EpochManagerReader<Addr = PeerAddress>>,
-        client_factory: TariValidatorNodeRpcClientFactory,
         substate_store: SqliteSubstateStore,
-        event_filters: Vec<EventFilterConfig>,
+        substate_scanner:
+            Arc<SubstateScanner<EpochManagerHandle<PeerAddress>, TariValidatorNodeRpcClientFactory, SubstateFileCache>>,
     ) -> Self {
         Self {
-            network,
-            epoch_manager,
-            client_factory,
             substate_store,
-            event_filters,
+            substate_scanner
         }
     }
 
-    pub async fn scan_events(&self) -> Result<usize, anyhow::Error> {
-        info!(
-            target: LOG_TARGET,
-            "scan_events",
-        );
+    pub fn save_event_to_db(
+        &self,
+        substate_id: &SubstateId,
+        template_address: TemplateAddress,
+        tx_hash: TransactionId,
+        topic: String,
+        payload: &Metadata,
+        version: u64,
+    ) -> Result<(), anyhow::Error> {
+        let mut tx = self.substate_store.create_write_tx()?;
+        let new_event = NewEvent {
+            substate_id: Some(substate_id.to_string()),
+            template_address: template_address.to_string(),
+            tx_hash: tx_hash.to_string(),
+            topic,
+            payload: payload.to_json().expect("Failed to convert to JSON"),
+            version: version as i32,
+        };
+        tx.save_event(new_event)?;
+        tx.commit()?;
+        Ok(())
+    }
 
-        let mut event_count = 0;
+    pub async fn scan_events_for_transaction(&self, tx_id: TransactionId) -> Result<Vec<Event>, anyhow::Error> {
+        let events = {
+            let mut tx = self.substate_store.create_read_tx()?;
+            tx.get_events_for_transaction(tx_id)?
+        };
 
-        let current_epoch = self.epoch_manager.current_epoch().await?;
-        // let network_committee_info = self.epoch_manager.get_network_committees().await?;
-        // let epoch = network_committee_info.epoch;
-        let current_committees = self.epoch_manager.get_committees(current_epoch).await?;
-        for (shard, committee) in current_committees {
-            info!(
-                target: LOG_TARGET,
-                "Scanning committee epoch={}, shard={}",
-                current_epoch,
-                shard
-            );
-            // TODO: use the latest block id that we scanned for each committee
-            let new_blocks = self
-                .get_new_blocks_from_committee(shard, &mut committee.clone(), current_epoch)
+        let mut events = events
+            .iter()
+            .map(|e| Event::try_from(e.clone()))
+            .collect::<Result<Vec<Event>, anyhow::Error>>()?;
+
+        // If we have no events locally, fetch from the network if possible.
+        if events.is_empty() {
+            let network_events = self.substate_scanner.get_events_for_transaction(tx_id).await?;
+            events.extend(network_events);
+        }
+
+        Ok(events)
+    }
+
+    pub async fn scan_events_for_substate_from_network(
+        &self,
+        substate_id: SubstateId,
+        version: Option<u32>,
+    ) -> Result<Vec<Event>, anyhow::Error> {
+        let mut events = vec![];
+        let version = version.unwrap_or_default();
+
+        // check if database contains the events for this transaction, by querying
+        // what is the latest version for the given component_address
+        let stored_versions_in_db;
+        {
+            let mut tx = self.substate_store.create_read_tx()?;
+            stored_versions_in_db = tx.get_stored_versions_of_events(&substate_id, version)?;
+
+            let stored_events = match tx.get_all_events(&substate_id) {
+                Ok(events) => events,
+                Err(e) => {
+                    info!(
+                        target: LOG_TARGET,
+                        "Failed to get all events for substate_id = {}, version = {} with error = {}",
+                        substate_id,
+                        version,
+                        e
+                    );
+                    return Err(e.into());
+                },
+            };
+
+            let stored_events = stored_events
+                .iter()
+                .map(|e| e.clone().try_into())
+                .collect::<Result<Vec<_>, _>>()?;
+            events.extend(stored_events);
+        }
+
+        for v in 0..version {
+            if stored_versions_in_db.contains(&v) {
+                continue;
+            }
+            let network_version_events = self
+                .substate_scanner
+                .get_events_for_substate_and_version(&substate_id, v)
                 .await?;
-            info!(
-                target: LOG_TARGET,
-                "Scanned {} blocks",
-                new_blocks.len()
-            );
-            let transaction_ids = self.extract_transaction_ids_from_blocks(new_blocks);
-            info!(
-                target: LOG_TARGET,
-                "Scanned {} transactions",
-                transaction_ids.len()
-            );
-
-            for transaction_id in transaction_ids {
-                // fetch all the events in the transaction
-                let events = self.get_events_for_transaction(transaction_id).await?;
-                event_count += events.len();
-
-                // only keep the events specified by the indexer filter
-                let filtered_events = events.into_iter().filter(|ev| self.should_persist_event(ev)).collect();
-                self.store_events_in_db(&filtered_events).await?;
-            }
+            events.extend(network_version_events);
         }
 
-        info!(
-            target: LOG_TARGET,
-            "Scanned {} events",
-            event_count
-        );
+        let latest_version_in_db = stored_versions_in_db.into_iter().max().unwrap_or_default();
+        let version = version.max(latest_version_in_db);
 
-        Ok(event_count)
+        // check if there are newest events for this component address in the network
+        let network_events = self
+            .substate_scanner
+            .get_events_for_substate(&substate_id, Some(version))
+            .await?;
+
+        // stores the newest network events to the db
+        // because the same substate_id with different version
+        // can be processed in the same transaction, we need to avoid
+        // duplicates
+        for (version, event) in network_events {
+            let template_address = event.template_address();
+            let tx_hash = TransactionId::new(event.tx_hash().into_array());
+            let topic = event.topic();
+            let payload = event.payload();
+            self.save_event_to_db(
+                &substate_id,
+                template_address,
+                tx_hash,
+                topic,
+                payload,
+                u64::from(version),
+            )?;
+            events.push(event);
+        }
+
+        Ok(events)
     }
 
-    fn should_persist_event(&self, event_data: &EventData) -> bool {
-        for filter in &self.event_filters {
-            if Self::event_matches_filter(filter, &event_data.event) {
-                return true;
-            }
-        }
-        
-        return false;
-    }
-
-    fn event_matches_filter(filter: &EventFilterConfig, event: &Event) -> bool {
-        let matches_topic = filter.topic.as_ref().map_or(true, |t| *t == event.topic());
-        let matches_template = filter.template_address.as_ref().map_or(true, |t| *t == event.template_address().to_string());
-
-        let matches_substate_id = match &filter.substate_id {
-            Some(substate_id) => event.substate_id().map(|s| s.to_string() == *substate_id).unwrap_or(false),
-            None => true,
+    pub async fn scan_events_by_payload(
+        &self,
+        payload_key: String,
+        payload_value: String,
+        offset: u32,
+        limit: u32,
+    ) -> Result<Vec<Event>, anyhow::Error> {
+        let events = {
+            let mut tx = self.substate_store.create_read_tx()?;
+            tx.get_events_by_payload(payload_key, payload_value, offset, limit)?
         };
 
-        let matches_entity_id = match &filter.entity_id {
-            Some(entity_id) => event.substate_id().map(|s| s.to_string().starts_with(entity_id)).unwrap_or(false),
-            None => true,
-        };
+        let events = events
+            .iter()
+            .map(|e| Event::try_from(e.clone()))
+            .collect::<Result<Vec<Event>, anyhow::Error>>()?;
 
-        if matches_topic && matches_template && matches_substate_id && matches_entity_id {
-            return true;
-        }
-
-        return false;
+        Ok(events)
     }
 
     pub async fn get_events_from_db(
@@ -185,330 +230,5 @@ impl EventManager {
         }
 
         Ok(events)
-    }
-
-    async fn store_events_in_db(&self, events_data: &Vec<EventData>) -> Result<(), anyhow::Error> {
-        let mut tx = self.substate_store.create_write_tx()?;
-
-        for data in events_data {
-            let event_row = NewEvent {
-                template_address: data.event.template_address().to_string(),
-                tx_hash: data.event.tx_hash().to_string(),
-                topic: data.event.topic(),
-                payload: data.event.payload().to_json().expect("Failed to convert to JSON"),
-                substate_id: data.event.substate_id().map(|s| s.to_string()),
-                version: 0_i32,
-            };
-
-            // TODO: properly avoid or handle duplicated events
-            //       For now we will just check if a similar event exists in the db
-            let event_already_exists = tx.event_exists(event_row.clone())?;
-            if event_already_exists {
-                // the event was already stored previously
-                warn!(
-                    target: LOG_TARGET,
-                    "Duplicated event {:}",
-                    data.event
-                );
-                continue;
-            }
-
-            tx.save_event(event_row)?;
-
-            // store/update the related substate if any
-            if let (Some(substate_id), Some(substate)) = (data.event.substate_id(), &data.substate) {
-                let substate_row = NewSubstate {
-                    address: substate_id.to_string(),
-                    version: i64::from(substate.version()),
-                    data: Self::encode_substate(&substate)?,
-                };
-                tx.set_substate(substate_row)?;
-            }
-        }
-
-        tx.commit()?;
-
-        Ok(())
-    }
-
-    fn encode_substate(substate: &Substate) -> Result<String, anyhow::Error> {
-        let pretty_json = serde_json::to_string_pretty(&substate)?;
-        Ok(pretty_json)
-    }
-
-    async fn get_events_for_transaction(&self, transaction_id: TransactionId) -> Result<Vec<EventData>, anyhow::Error> {
-        let committee = self.get_all_vns().await?;
-
-        for member in &committee {
-            let resp = self.get_execute_result_from_vn(member, &transaction_id).await;
-
-            match resp {
-                Ok(res) => {
-                    if let Some(execute_result) = res {
-                        let events = self.extract_events_from_transaction_result(execute_result);
-                        return Ok(events);
-                    } else {
-                        // The transaction is not successful, so we don't return any events
-                        return Ok(vec![]);
-                    }
-                },
-                Err(e) => {
-                    // We do nothing on a single VN failure, we only log it
-                    warn!(
-                        target: LOG_TARGET,
-                        "Could not get transaction result from vn {}: {}",
-                        member,
-                        e
-                    );
-                },
-            };
-        }
-
-        warn!(
-            target: LOG_TARGET,
-            "We could not get transaction results from any of the vns",
-        );
-        Ok(vec![])
-    }
-
-    async fn get_execute_result_from_vn(
-        &self,
-        vn_addr: &PeerAddress,
-        transaction_id: &TransactionId,
-    ) -> Result<Option<ExecuteResult>, anyhow::Error> {
-        let mut rpc_client = self.client_factory.create_client(vn_addr);
-        let mut client = rpc_client.client_connection().await?;
-
-        let response = client
-            .get_transaction_result(GetTransactionResultRequest {
-                transaction_id: transaction_id.as_bytes().to_vec(),
-            })
-            .await?;
-
-        match PayloadResultStatus::try_from(response.status) {
-            Ok(PayloadResultStatus::Finalized) => {
-                let proto_decision = tari_dan_p2p::proto::consensus::Decision::try_from(response.final_decision)?;
-                let final_decision = proto_decision.try_into()?;
-                if let Decision::Commit = final_decision {
-                    Ok(Some(response.execution_result)
-                        .filter(|r| !r.is_empty())
-                        .map(|r| decode(&r))
-                        .transpose()?)
-                } else {
-                    Ok(None)
-                }
-            },
-            _ => Ok(None),
-        }
-    }
-
-    fn extract_events_from_transaction_result(&self, result: ExecuteResult) -> Vec<EventData> {
-        if let TransactionResult::Accept(substate_diff) = result.finalize.result {
-            let substates: HashMap<SubstateId, Substate> = substate_diff.into_up_iter().collect();
-
-            let events = result.finalize.events
-                .into_iter()
-                .map(|event| {
-                    let substate = if let Some(substate_id) = event.substate_id() {
-                        substates.get(&substate_id).cloned()
-                    } else {
-                        None
-                    };
-
-                    EventData {
-                        event,
-                        substate,
-                    }
-                }).collect();
-
-            return events;
-        } else {
-            return vec![]
-        }
-    }
-
-    fn extract_transaction_ids_from_blocks(&self, blocks: Vec<Block>) -> HashSet<TransactionId> {
-        let mut transaction_ids = HashSet::new();
-
-        for block in blocks {
-            for command in block.commands() {
-                match command {
-                    Command::Accept(t) | Command::LocalOnly(t) => {
-                        transaction_ids.insert(*t.id());
-                    },
-                    _ => {
-                        // we are only interested in events from confirmed transactions
-                    },
-                }
-            }
-        }
-
-        transaction_ids
-    }
-
-    fn build_genesis_block_id(&self) -> BlockId {
-        let start_block = Block::zero_block(self.network);
-        *start_block.id()
-    }
-
-    #[allow(unused_assignments)]
-    async fn get_new_blocks_from_committee(
-        &self,
-        shard: Shard,
-        committee: &mut Committee<PeerAddress>,
-        epoch: Epoch,
-    ) -> Result<Vec<Block>, anyhow::Error> {
-        // We start scanning from the last scanned block for this commitee
-        let start_block_id = {
-            let mut tx = self.substate_store.create_read_tx()?;
-            tx.get_last_scanned_block_id(epoch, shard)?
-        };
-        let start_block_id = start_block_id.unwrap_or(self.build_genesis_block_id());
-
-        committee.members.shuffle(&mut OsRng);
-        let mut last_block_id = start_block_id;
-
-        info!(
-            target: LOG_TARGET,
-            "Scanning new blocks since {} from (epoch={}, shard={})",
-            last_block_id,
-            epoch,
-            shard
-        );
-
-        for (member, _) in &committee.members {
-            debug!(
-                target: LOG_TARGET,
-                "Trying to get blocks from VN {} (epoch={}, shard={})",
-                member,
-                epoch,
-                shard
-            );
-            let resp = self.get_blocks_from_vn(member, start_block_id).await;
-
-            match resp {
-                Ok(blocks) => {
-                    // TODO: try more than 1 VN per commitee
-                    info!(
-                        target: LOG_TARGET,
-                        "Got {} blocks from VN {} (epoch={}, shard={})",
-                        blocks.len(),
-                        member,
-                        epoch,
-                        shard,
-                    );
-                    if let Some(block) = blocks.last() {
-                        last_block_id = *block.id();
-                    }
-                    // Store the latest scanned block id in the database for future scans
-                    self.save_scanned_block_id(epoch, shard, last_block_id)?;
-                    return Ok(blocks);
-                },
-                Err(e) => {
-                    // We do nothing on a single VN failure, we only log it
-                    warn!(
-                        target: LOG_TARGET,
-                        "Could not get blocks from VN {} (epoch={}, shard={}): {}",
-                        member,
-                        epoch,
-                        shard,
-                        e
-                    );
-                },
-            };
-        }
-
-        // We don't raise an error if none of the VNs have blocks, the scanning will retry eventually
-        warn!(
-            target: LOG_TARGET,
-            "Could not get blocks from any of the VNs of the committee (epoch={}, shard={})",
-            epoch,
-            shard
-        );
-        Ok(vec![])
-    }
-
-    fn save_scanned_block_id(&self, epoch: Epoch, shard: Shard, last_block_id: BlockId) -> Result<(), anyhow::Error> {
-        let row = NewScannedBlockId {
-            epoch: epoch.0 as i64,
-            shard: i64::from(shard.as_u32()),
-            last_block_id: last_block_id.as_bytes().to_vec(),
-        };
-        let mut tx = self.substate_store.create_write_tx()?;
-        tx.save_scanned_block_id(row)?;
-        tx.commit()?;
-        Ok(())
-    }
-
-    async fn get_all_vns(&self) -> Result<Vec<PeerAddress>, anyhow::Error> {
-        // get all the committees
-        let epoch = self.epoch_manager.current_epoch().await?;
-        Ok(self
-            .epoch_manager
-            .get_all_validator_nodes(epoch)
-            .await
-            .map(|v| v.iter().map(|m| m.address).collect())?)
-    }
-
-    async fn get_blocks_from_vn(
-        &self,
-        vn_addr: &PeerAddress,
-        start_block_id: BlockId,
-    ) -> Result<Vec<Block>, anyhow::Error> {
-        let mut blocks = vec![];
-
-        let mut rpc_client = self.client_factory.create_client(vn_addr);
-        let mut client = rpc_client.client_connection().await?;
-
-        let mut stream = client
-            .sync_blocks(SyncBlocksRequest {
-                start_block_id: start_block_id.as_bytes().to_vec(),
-                up_to_epoch: None,
-            })
-            .await?;
-        while let Some(resp) = stream.next().await {
-            let msg = resp?;
-
-            let new_block = msg
-                .into_block()
-                .ok_or_else(|| anyhow::anyhow!("Expected peer to return a newblock"))?;
-            let block = Block::try_from(new_block)?;
-
-            let Some(_) = stream.next().await else {
-                anyhow::bail!("Peer closed session before sending QC message")
-            };
-
-            let Some(resp) = stream.next().await else {
-                anyhow::bail!("Peer closed session before sending substate update count message")
-            };
-            let msg = resp?;
-            let num_substates =
-                msg.substate_count()
-                    .ok_or_else(|| anyhow::anyhow!("Expected peer to return substate count"))? as usize;
-
-            for _ in 0..num_substates {
-                let Some(_) = stream.next().await else {
-                    anyhow::bail!("Peer closed session before sending substate updates message")
-                };
-            }
-
-            let Some(resp) = stream.next().await else {
-                anyhow::bail!("Peer closed session before sending transactions message")
-            };
-            let msg = resp?;
-            let transactions = msg
-                .into_transactions()
-                .ok_or_else(|| anyhow::anyhow!("Expected peer to return transactions"))?;
-
-            let _transactions = transactions
-                .into_iter()
-                .map(Transaction::try_from)
-                .map(|r| r.map(TransactionRecord::new))
-                .collect::<Result<Vec<_>, _>>()?;
-
-            blocks.push(block);
-        }
-
-        Ok(blocks)
     }
 }
